+++ .github/workflows/build.yml
@@ -25,7 +25,6 @@ jobs:
   phpunit:
     uses: yiisoft/actions/.github/workflows/phpunit.yml@master
     with:
-      extensions: :intl
       coverage: xdebug
       os: >-
         ['ubuntu-latest', 'windows-latest']
@@ -45,6 +44,16 @@ jobs:
     with:
       phpunitConfig: phpunit-with-intl.xml
       extensions: intl
+      coverage: xdebug
+      os: >-
+        ['ubuntu-latest', 'windows-latest']
+      php: >-
+        ['8.0', '8.1']
+  phpunit-without-intl:
+    uses: yiisoft/actions/.github/workflows/phpunit.yml@master
+    with:
+      phpunitConfig: phpunit-without-intl.xml
+      extensions: :intl
       coverage: xdebug
       os: >-
         ['ubuntu-latest', 'windows-latest']
