+++ phpunit-without-intl.xml.dist
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<phpunit colors="true"
+         verbose="true"
+         failOnRisky="true"
+         failOnWarning="true"
+         convertErrorsToExceptions="true"
+         convertNoticesToExceptions="true"
+         convertWarningsToExceptions="true"
+         stopOnFailure="false"
+         executionOrder="random"
+         resolveDependencies="true">
+    <php>
+        <ini name="error_reporting" value="-1"/>
+    </php>
+
+    <testsuites>
+        <testsuite name="Validator">
+            <directory>./tests/TestEnvironments/WithoutIntl</directory>
+        </testsuite>
+    </testsuites>
+    <coverage>
+        <include>
+            <directory>./src</directory>
+            <directory>./config</directory>
+        </include>
+    </coverage>
+</phpunit>
