+++ tests/Rule/UrlTest.php
@@ -5,14 +5,11 @@ declare(strict_types=1);
 namespace Yiisoft\Validator\Tests\Rule;
 
 use PHPUnit\Framework\TestCase;
-use RuntimeException;
 use Yiisoft\Validator\Rule\Url;
 use Yiisoft\Validator\Rule\UrlHandler;
 use Yiisoft\Validator\Tests\Rule\Base\DifferentRuleInHandlerTestTrait;
 use Yiisoft\Validator\Tests\Rule\Base\SkipOnErrorTestTrait;
 use Yiisoft\Validator\Tests\Rule\Base\WhenTestTrait;
-
-use function extension_loaded;
 
 final class UrlTest extends TestCase
 {
@@ -32,16 +29,6 @@ final class UrlTest extends TestCase
     {
         $rule = new Url(validSchemes: ['http', 'https', 'ftp', 'ftps']);
         $this->assertSame(['http', 'https', 'ftp', 'ftps'], $rule->getValidSchemes());
-    }
-
-    public function testEnableIdnWithMissingIntlExtension(): void
-    {
-        if (extension_loaded('intl')) {
-            $this->markTestSkipped('The intl extension must be unavailable for this test.');
-        }
-
-        $this->expectException(RuntimeException::class);
-        new Url(enableIDN: true);
     }
 
     public function testSkipOnError(): void
